Adjust file organization for new Jabbeke files.
In the new files the timestamp in the filename no longer corresponds to
the start of scanning, but to the end of processing, roughly 5 minutes
later. The start time can only be retrieved from inspection of the file.
Therefore, the start time is now estimated from the name to determine
which files are not yet retrieved. After retrieving the file, the correct
start time is read from the contents during the organize step.